[Silver II] Title: 나무 자르기, Time: 476 ms, Memory: 119456 KB -BaekjoonHub

+++ 백준/Silver/2805. 나무 자르기/나무 자르기.java
@@ -36,10 +36,10 @@ public class Main {
 			}
 			
 			if (s >= M) {
-				result = (int) mid;
-				min = (int) mid + 1;
+				result = mid;
+				min = mid + 1;
 			} else if (s < M) {
-				max = (int) mid - 1;
+				max = mid - 1;
 			}
 
 		}
